refactor and implement resizing tests
